+++ src/main/java/net/sashiro/additionalvanillastuff/data/generators/GenLootTableProvider.java
@@ -1,58 +1,43 @@
 package net.sashiro.additionalvanillastuff.data.generators;
-/*
-import com.google.common.collect.ImmutableList;
-import com.mojang.datafixers.util.Pair;
-import net.minecraft.data.DataGenerator;
-import net.minecraft.data.loot.BlockLoot;
+
+import net.minecraft.data.PackOutput;
+import net.minecraft.data.loot.BlockLootSubProvider;
 import net.minecraft.data.loot.LootTableProvider;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.flag.FeatureFlags;
 import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.storage.loot.LootTable;
-import net.minecraft.world.level.storage.loot.LootTables;
-import net.minecraft.world.level.storage.loot.ValidationContext;
-import net.minecraft.world.level.storage.loot.parameters.LootContextParamSet;
 import net.minecraft.world.level.storage.loot.parameters.LootContextParamSets;
 import net.minecraftforge.registries.RegistryObject;
 import net.sashiro.additionalvanillastuff.event.ModRegistryEvent;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
-import java.util.Map;
-import java.util.function.BiConsumer;
-import java.util.function.Consumer;
-import java.util.function.Supplier;
+import java.util.Set;
 import java.util.stream.Collectors;
 
-public class GenLootTableProvider extends LootTableProvider {
-    public GenLootTableProvider(DataGenerator dataGeneratorIn) {
-        super(dataGeneratorIn);
+public class GenLootTableProvider {
+
+    public static LootTableProvider create(PackOutput output) {
+        return new LootTableProvider(output, Set.of(),
+                List.of(new LootTableProvider.SubProviderEntry(BlocksLootTable::new, LootContextParamSets.BLOCK)));
     }
 
-    @Override
-    protected List<Pair<Supplier<Consumer<BiConsumer<ResourceLocation, LootTable.Builder>>>, LootContextParamSet>> getTables() {
-        return ImmutableList.of(
-                Pair.of(CompressedLootTable::new, LootContextParamSets.BLOCK)
-        );
-    }
+    public static class BlocksLootTable extends BlockLootSubProvider {
+        protected BlocksLootTable() {
+            super(Set.of(), FeatureFlags.REGISTRY.allFlags());
+        }
 
-    @Override
-    protected void validate(Map<ResourceLocation, LootTable> map, ValidationContext validationtracker) {
-        map.forEach((id, lootTable) -> LootTables.validate(validationtracker, id, lootTable));
-    }
-
-    public static class CompressedLootTable extends BlockLoot {
         @Override
-        protected void addTables() {
+        protected void generate() {
             for (RegistryObject<Block> block : ModRegistryEvent.BLOCK_REGISTRY.getEntries()) {
                 dropSelf(block.get());
             }
         }
 
         @Override
-        protected Iterable<Block> getKnownBlocks() {
+        protected @NotNull Iterable<Block> getKnownBlocks() {
             return ModRegistryEvent.BLOCK_REGISTRY.getEntries().stream()
                     .map(RegistryObject::get)
                     .collect(Collectors.toList());
         }
     }
 }
- */