+++ src/main/java/net/sashiro/additionalvanillastuff/data/generators/GenRecipeProvider.java
@@ -1,12 +1,12 @@
 package net.sashiro.additionalvanillastuff.data.generators;
-/*
-import net.minecraft.data.DataGenerator;
+
+import net.minecraft.data.PackOutput;
 import net.minecraft.data.recipes.FinishedRecipe;
+import net.minecraft.data.recipes.RecipeCategory;
 import net.minecraft.data.recipes.RecipeProvider;
 import net.minecraft.data.recipes.ShapedRecipeBuilder;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.Blocks;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
 import net.sashiro.additionalvanillastuff.AdditionalVanillaStuff;
@@ -15,56 +15,29 @@ import net.sashiro.additionalvanillastuff.event.ModRegistryEvent;
 import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.function.Consumer;
 
+@SuppressWarnings("DuplicateCondition")
 public class GenRecipeProvider extends RecipeProvider {
-    public GenRecipeProvider(DataGenerator generatorIn) {
-        super(generatorIn);
+    public GenRecipeProvider(PackOutput packOutput) {
+        super(packOutput);
     }
 
     @Override
     @ParametersAreNonnullByDefault
-    protected void buildCraftingRecipes(Consumer<FinishedRecipe> consumer) {
+    protected void buildRecipes(Consumer<FinishedRecipe> consumer) {
 
         for (RegistryObject<Block> block : ModRegistryEvent.BLOCK_REGISTRY.getEntries()) {
-            String loc = block.get().getRegistryName().getPath();
-            Block vanillaBLock = Blocks.AIR;
+            String loc = block.get().getDescriptionId().replace("block.additionalvanillastuff.", "");
+            Block vanillaBLock;
             for (Block mcBlock : ForgeRegistries.BLOCKS) {
-                String blockName = block.get().getRegistryName().getPath().replace("_wall", "").replace("_stairs", "").replace("_slab", "");
-                String mcBlockName = mcBlock.getRegistryName().getPath();
+                String blockName = loc.replace("_wall", "").replace("_stairs", "").replace("_slab", "");
+                String mcBlockName = mcBlock.getDescriptionId().replace("block.minecraft.", "");
 
-                if (mcBlockName.equals(blockName))  {
+                if (mcBlockName.equals(blockName)) {
                     vanillaBLock = mcBlock;
                     if (!loc.contains("wool")) {
                         if (loc.contains("stairs")) {
-                            stonecutterResultFromBase(consumer, block.get(), vanillaBLock);
-                            ShapedRecipeBuilder.shaped(block.get(), 4) // result
-                                    .define('#', vanillaBLock) // ingredient
-                                    .pattern("#  ")
-                                    .pattern("## ")
-                                    .pattern("###")
-                                    .unlockedBy("has_item", has(vanillaBLock.asItem()))
-                                    .save(consumer, new ResourceLocation(AdditionalVanillaStuff.MOD_ID, "shaped_" + loc));
-                        }
-                        else if (loc.contains("slab")) {
-                            stonecutterResultFromBase(consumer, block.get(), vanillaBLock, 2);
-                            ShapedRecipeBuilder.shaped(block.get(), 6) // result
-                                    .define('#', vanillaBLock) // ingredient
-                                    .pattern("###")
-                                    .unlockedBy("has_item", has(vanillaBLock.asItem()))
-                                    .save(consumer, new ResourceLocation(AdditionalVanillaStuff.MOD_ID, "shaped_" + loc));
-                        }
-                        else if (loc.contains("wall")) {
-                            stonecutterResultFromBase(consumer, block.get(), vanillaBLock);
-                            ShapedRecipeBuilder.shaped(block.get(), 6) // result
-                                    .define('#', vanillaBLock) // ingredient
-                                    .pattern("###")
-                                    .pattern("###")
-                                    .unlockedBy("has_item", has(vanillaBLock.asItem()))
-                                    .save(consumer, new ResourceLocation(AdditionalVanillaStuff.MOD_ID, "shaped_" + loc));
-                        }
-                    }
-                    else if (loc.contains("wool")) {
-                        if (loc.contains("stairs")) {
-                            ShapedRecipeBuilder.shaped(block.get(), 4) // result
+                            stonecutterResultFromBase(consumer, RecipeCategory.BUILDING_BLOCKS, block.get(), vanillaBLock);
+                            ShapedRecipeBuilder.shaped(RecipeCategory.BUILDING_BLOCKS, block.get(), 4) // result
                                     .define('#', vanillaBLock) // ingredient
                                     .pattern("#  ")
                                     .pattern("## ")
@@ -72,13 +45,38 @@ public class GenRecipeProvider extends RecipeProvider {
                                     .unlockedBy("has_item", has(vanillaBLock.asItem()))
                                     .save(consumer, new ResourceLocation(AdditionalVanillaStuff.MOD_ID, "shaped_" + loc));
                         } else if (loc.contains("slab")) {
-                            ShapedRecipeBuilder.shaped(block.get(), 6) // result
+                            stonecutterResultFromBase(consumer, RecipeCategory.BUILDING_BLOCKS, block.get(), vanillaBLock, 2);
+                            ShapedRecipeBuilder.shaped(RecipeCategory.BUILDING_BLOCKS, block.get(), 6) // result
                                     .define('#', vanillaBLock) // ingredient
                                     .pattern("###")
                                     .unlockedBy("has_item", has(vanillaBLock.asItem()))
                                     .save(consumer, new ResourceLocation(AdditionalVanillaStuff.MOD_ID, "shaped_" + loc));
                         } else if (loc.contains("wall")) {
-                            ShapedRecipeBuilder.shaped(block.get(), 6) // result
+                            stonecutterResultFromBase(consumer, RecipeCategory.BUILDING_BLOCKS, block.get(), vanillaBLock);
+                            ShapedRecipeBuilder.shaped(RecipeCategory.BUILDING_BLOCKS, block.get(), 6) // result
+                                    .define('#', vanillaBLock) // ingredient
+                                    .pattern("###")
+                                    .pattern("###")
+                                    .unlockedBy("has_item", has(vanillaBLock.asItem()))
+                                    .save(consumer, new ResourceLocation(AdditionalVanillaStuff.MOD_ID, "shaped_" + loc));
+                        }
+                    } else if (loc.contains("wool")) {
+                        if (loc.contains("stairs")) {
+                            ShapedRecipeBuilder.shaped(RecipeCategory.BUILDING_BLOCKS, block.get(), 4) // result
+                                    .define('#', vanillaBLock) // ingredient
+                                    .pattern("#  ")
+                                    .pattern("## ")
+                                    .pattern("###")
+                                    .unlockedBy("has_item", has(vanillaBLock.asItem()))
+                                    .save(consumer, new ResourceLocation(AdditionalVanillaStuff.MOD_ID, "shaped_" + loc));
+                        } else if (loc.contains("slab")) {
+                            ShapedRecipeBuilder.shaped(RecipeCategory.BUILDING_BLOCKS, block.get(), 6) // result
+                                    .define('#', vanillaBLock) // ingredient
+                                    .pattern("###")
+                                    .unlockedBy("has_item", has(vanillaBLock.asItem()))
+                                    .save(consumer, new ResourceLocation(AdditionalVanillaStuff.MOD_ID, "shaped_" + loc));
+                        } else if (loc.contains("wall")) {
+                            ShapedRecipeBuilder.shaped(RecipeCategory.BUILDING_BLOCKS, block.get(), 6) // result
                                     .define('#', vanillaBLock) // ingredient
                                     .pattern("###")
                                     .pattern("###")
@@ -98,7 +96,6 @@ public class GenRecipeProvider extends RecipeProvider {
                             .pattern("###")
                             .unlockedBy("has_item", has(blocks.get(i - 1).get().asItem()))
                             .save(consumer, new ResourceLocation("compressedblocks", "shaped_" + blockName.replace("compressedblocks:", "")));
-/
+*/
     }
 }
-*/