+++ src/main/java/net/sashiro/additionalvanillastuff/world/level/item/ModItems.java
@@ -2,7 +2,6 @@ package net.sashiro.additionalvanillastuff.world.level.item;
 
 
 import net.minecraft.world.item.BlockItem;
-import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.Item;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.registries.RegistryObject;
@@ -11,10 +10,14 @@ import net.sashiro.additionalvanillastuff.world.level.block.ModBlocks;
 
 import static net.sashiro.additionalvanillastuff.event.ModRegistryEvent.ITEM_REGISTRY;
 
+@SuppressWarnings("unused")
 @Mod.EventBusSubscriber(modid = AdditionalVanillaStuff.MOD_ID)
 public class ModItems {
     private static final Item.Properties PROPERTIES = new Item.Properties();
 
+    // DIRT STAIRS
+    public static final RegistryObject<Item> DIRT_STAIRS = ITEM_REGISTRY.register("dirt_stairs", () -> new BlockItem(ModBlocks.DIRT_STAIRS.get(), PROPERTIES));
+    // TERRACOTTA STAIRS
     public static final RegistryObject<Item> WHITE_TERRACOTTA_STAIRS = ITEM_REGISTRY.register("white_terracotta_stairs", () -> new BlockItem(ModBlocks.WHITE_TERRACOTTA_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> ORANGE_TERRACOTTA_STAIRS = ITEM_REGISTRY.register("orange_terracotta_stairs", () -> new BlockItem(ModBlocks.ORANGE_TERRACOTTA_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> MAGENTA_TERRACOTTA_STAIRS = ITEM_REGISTRY.register("magenta_terracotta_stairs", () -> new BlockItem(ModBlocks.MAGENTA_TERRACOTTA_STAIRS.get(), PROPERTIES));
@@ -48,6 +51,7 @@ public class ModItems {
     public static final RegistryObject<Item> GREEN_GLAZED_TERRACOTTA_STAIRS = ITEM_REGISTRY.register("green_glazed_terracotta_stairs", () -> new BlockItem(ModBlocks.GREEN_GLAZED_TERRACOTTA_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> RED_GLAZED_TERRACOTTA_STAIRS = ITEM_REGISTRY.register("red_glazed_terracotta_stairs", () -> new BlockItem(ModBlocks.RED_GLAZED_TERRACOTTA_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> BLACK_GLAZED_TERRACOTTA_STAIRS = ITEM_REGISTRY.register("black_glazed_terracotta_stairs", () -> new BlockItem(ModBlocks.BLACK_GLAZED_TERRACOTTA_STAIRS.get(), PROPERTIES));
+    // CONCRETE STAIRS
     public static final RegistryObject<Item> WHITE_CONCRETE_STAIRS = ITEM_REGISTRY.register("white_concrete_stairs", () -> new BlockItem(ModBlocks.WHITE_CONCRETE_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> ORANGE_CONCRETE_STAIRS = ITEM_REGISTRY.register("orange_concrete_stairs", () -> new BlockItem(ModBlocks.ORANGE_CONCRETE_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> MAGENTA_CONCRETE_STAIRS = ITEM_REGISTRY.register("magenta_concrete_stairs", () -> new BlockItem(ModBlocks.MAGENTA_CONCRETE_STAIRS.get(), PROPERTIES));
@@ -64,6 +68,7 @@ public class ModItems {
     public static final RegistryObject<Item> GREEN_CONCRETE_STAIRS = ITEM_REGISTRY.register("green_concrete_stairs", () -> new BlockItem(ModBlocks.GREEN_CONCRETE_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> RED_CONCRETE_STAIRS = ITEM_REGISTRY.register("red_concrete_stairs", () -> new BlockItem(ModBlocks.RED_CONCRETE_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> BLACK_CONCRETE_STAIRS = ITEM_REGISTRY.register("black_concrete_stairs", () -> new BlockItem(ModBlocks.BLACK_CONCRETE_STAIRS.get(), PROPERTIES));
+    // WOOL STAIRS
     public static final RegistryObject<Item> WHITE_WOOL_STAIRS = ITEM_REGISTRY.register("white_wool_stairs", () -> new BlockItem(ModBlocks.WHITE_WOOL_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> ORANGE_WOOL_STAIRS = ITEM_REGISTRY.register("orange_wool_stairs", () -> new BlockItem(ModBlocks.ORANGE_WOOL_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> MAGENTA_WOOL_STAIRS = ITEM_REGISTRY.register("magenta_wool_stairs", () -> new BlockItem(ModBlocks.MAGENTA_WOOL_STAIRS.get(), PROPERTIES));
@@ -80,6 +85,7 @@ public class ModItems {
     public static final RegistryObject<Item> GREEN_WOOL_STAIRS = ITEM_REGISTRY.register("green_wool_stairs", () -> new BlockItem(ModBlocks.GREEN_WOOL_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> RED_WOOL_STAIRS = ITEM_REGISTRY.register("red_wool_stairs", () -> new BlockItem(ModBlocks.RED_WOOL_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> BLACK_WOOL_STAIRS = ITEM_REGISTRY.register("black_wool_stairs", () -> new BlockItem(ModBlocks.BLACK_WOOL_STAIRS.get(), PROPERTIES));
+    // METAL STAIRS
     public static final RegistryObject<Item> IRON_BLOCK_STAIRS = ITEM_REGISTRY.register("iron_block_stairs", () -> new BlockItem(ModBlocks.IRON_BLOCK_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> GOLD_BLOCK_STAIRS = ITEM_REGISTRY.register("gold_block_stairs", () -> new BlockItem(ModBlocks.GOLD_BLOCK_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> DIAMOND_BLOCK_STAIRS = ITEM_REGISTRY.register("diamond_block_stairs", () -> new BlockItem(ModBlocks.DIAMOND_BLOCK_STAIRS.get(), PROPERTIES));
@@ -87,11 +93,15 @@ public class ModItems {
     public static final RegistryObject<Item> NETHERITE_BLOCK_STAIRS = ITEM_REGISTRY.register("netherite_block_stairs", () -> new BlockItem(ModBlocks.NETHERITE_BLOCK_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> LAPIS_BLOCK_STAIRS = ITEM_REGISTRY.register("lapis_block_stairs", () -> new BlockItem(ModBlocks.LAPIS_BLOCK_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> EMERALD_BLOCK_STAIRS = ITEM_REGISTRY.register("emerald_block_stairs", () -> new BlockItem(ModBlocks.EMERALD_BLOCK_STAIRS.get(), PROPERTIES));
+    // STONE STAIRS
     public static final RegistryObject<Item> CALCITE_STAIRS = ITEM_REGISTRY.register("calcite_stairs", () -> new BlockItem(ModBlocks.CALCITE_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> TUFF_STAIRS = ITEM_REGISTRY.register("tuff_stairs", () -> new BlockItem(ModBlocks.TUFF_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> SMOOTH_BASALT_STAIRS = ITEM_REGISTRY.register("smooth_basalt_stairs", () -> new BlockItem(ModBlocks.SMOOTH_BASALT_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> OBSIDIAN_STAIRS = ITEM_REGISTRY.register("obsidian_stairs", () -> new BlockItem(ModBlocks.OBSIDIAN_STAIRS.get(), PROPERTIES));
     public static final RegistryObject<Item> CRYING_OBSIDIAN_STAIRS = ITEM_REGISTRY.register("crying_obsidian_stairs", () -> new BlockItem(ModBlocks.CRYING_OBSIDIAN_STAIRS.get(), PROPERTIES));
+    // DIRT SLABS
+    public static final RegistryObject<Item> DIRT_SLAB = ITEM_REGISTRY.register("dirt_slab", () -> new BlockItem(ModBlocks.DIRT_SLAB.get(), PROPERTIES));
+    // TERRACOTTA SLABS
     public static final RegistryObject<Item> WHITE_TERRACOTTA_SLAB = ITEM_REGISTRY.register("white_terracotta_slab", () -> new BlockItem(ModBlocks.WHITE_TERRACOTTA_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> ORANGE_TERRACOTTA_SLAB = ITEM_REGISTRY.register("orange_terracotta_slab", () -> new BlockItem(ModBlocks.ORANGE_TERRACOTTA_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> MAGENTA_TERRACOTTA_SLAB = ITEM_REGISTRY.register("magenta_terracotta_slab", () -> new BlockItem(ModBlocks.MAGENTA_TERRACOTTA_SLAB.get(), PROPERTIES));
@@ -125,6 +135,7 @@ public class ModItems {
     public static final RegistryObject<Item> GREEN_GLAZED_TERRACOTTA_SLAB = ITEM_REGISTRY.register("green_glazed_terracotta_slab", () -> new BlockItem(ModBlocks.GREEN_GLAZED_TERRACOTTA_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> RED_GLAZED_TERRACOTTA_SLAB = ITEM_REGISTRY.register("red_glazed_terracotta_slab", () -> new BlockItem(ModBlocks.RED_GLAZED_TERRACOTTA_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> BLACK_GLAZED_TERRACOTTA_SLAB = ITEM_REGISTRY.register("black_glazed_terracotta_slab", () -> new BlockItem(ModBlocks.BLACK_GLAZED_TERRACOTTA_SLAB.get(), PROPERTIES));
+    // CONCRETE SLABS
     public static final RegistryObject<Item> WHITE_CONCRETE_SLAB = ITEM_REGISTRY.register("white_concrete_slab", () -> new BlockItem(ModBlocks.WHITE_CONCRETE_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> ORANGE_CONCRETE_SLAB = ITEM_REGISTRY.register("orange_concrete_slab", () -> new BlockItem(ModBlocks.ORANGE_CONCRETE_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> MAGENTA_CONCRETE_SLAB = ITEM_REGISTRY.register("magenta_concrete_slab", () -> new BlockItem(ModBlocks.MAGENTA_CONCRETE_SLAB.get(), PROPERTIES));
@@ -141,6 +152,7 @@ public class ModItems {
     public static final RegistryObject<Item> GREEN_CONCRETE_SLAB = ITEM_REGISTRY.register("green_concrete_slab", () -> new BlockItem(ModBlocks.GREEN_CONCRETE_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> RED_CONCRETE_SLAB = ITEM_REGISTRY.register("red_concrete_slab", () -> new BlockItem(ModBlocks.RED_CONCRETE_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> BLACK_CONCRETE_SLAB = ITEM_REGISTRY.register("black_concrete_slab", () -> new BlockItem(ModBlocks.BLACK_CONCRETE_SLAB.get(), PROPERTIES));
+    // WOOL SLABS
     public static final RegistryObject<Item> WHITE_WOOL_SLAB = ITEM_REGISTRY.register("white_wool_slab", () -> new BlockItem(ModBlocks.WHITE_WOOL_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> ORANGE_WOOL_SLAB = ITEM_REGISTRY.register("orange_wool_slab", () -> new BlockItem(ModBlocks.ORANGE_WOOL_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> MAGENTA_WOOL_SLAB = ITEM_REGISTRY.register("magenta_wool_slab", () -> new BlockItem(ModBlocks.MAGENTA_WOOL_SLAB.get(), PROPERTIES));
@@ -157,6 +169,7 @@ public class ModItems {
     public static final RegistryObject<Item> GREEN_WOOL_SLAB = ITEM_REGISTRY.register("green_wool_slab", () -> new BlockItem(ModBlocks.GREEN_WOOL_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> RED_WOOL_SLAB = ITEM_REGISTRY.register("red_wool_slab", () -> new BlockItem(ModBlocks.RED_WOOL_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> BLACK_WOOL_SLAB = ITEM_REGISTRY.register("black_wool_slab", () -> new BlockItem(ModBlocks.BLACK_WOOL_SLAB.get(), PROPERTIES));
+    // METAL SLABS
     public static final RegistryObject<Item> IRON_BLOCK_SLAB = ITEM_REGISTRY.register("iron_block_slab", () -> new BlockItem(ModBlocks.IRON_BLOCK_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> GOLD_BLOCK_SLAB = ITEM_REGISTRY.register("gold_block_slab", () -> new BlockItem(ModBlocks.GOLD_BLOCK_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> DIAMOND_BLOCK_SLAB = ITEM_REGISTRY.register("diamond_block_slab", () -> new BlockItem(ModBlocks.DIAMOND_BLOCK_SLAB.get(), PROPERTIES));
@@ -164,11 +177,15 @@ public class ModItems {
     public static final RegistryObject<Item> NETHERITE_BLOCK_SLAB = ITEM_REGISTRY.register("netherite_block_slab", () -> new BlockItem(ModBlocks.NETHERITE_BLOCK_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> LAPIS_BLOCK_SLAB = ITEM_REGISTRY.register("lapis_block_slab", () -> new BlockItem(ModBlocks.LAPIS_BLOCK_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> EMERALD_BLOCK_SLAB = ITEM_REGISTRY.register("emerald_block_slab", () -> new BlockItem(ModBlocks.EMERALD_BLOCK_SLAB.get(), PROPERTIES));
+    // STONE SLABS
     public static final RegistryObject<Item> CALCITE_SLAB = ITEM_REGISTRY.register("calcite_slab", () -> new BlockItem(ModBlocks.CALCITE_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> TUFF_SLAB = ITEM_REGISTRY.register("tuff_slab", () -> new BlockItem(ModBlocks.TUFF_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> SMOOTH_BASALT_SLAB = ITEM_REGISTRY.register("smooth_basalt_slab", () -> new BlockItem(ModBlocks.SMOOTH_BASALT_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> OBSIDIAN_SLAB = ITEM_REGISTRY.register("obsidian_slab", () -> new BlockItem(ModBlocks.OBSIDIAN_SLAB.get(), PROPERTIES));
     public static final RegistryObject<Item> CRYING_OBSIDIAN_SLAB = ITEM_REGISTRY.register("crying_obsidian_slab", () -> new BlockItem(ModBlocks.CRYING_OBSIDIAN_SLAB.get(), PROPERTIES));
+    // DIRT STAIRS
+    public static final RegistryObject<Item> DIRT_WALL = ITEM_REGISTRY.register("dirt_wall", () -> new BlockItem(ModBlocks.DIRT_WALL.get(), PROPERTIES));
+    // TERRACOTTA WALLS
     public static final RegistryObject<Item> WHITE_TERRACOTTA_WALL = ITEM_REGISTRY.register("white_terracotta_wall", () -> new BlockItem(ModBlocks.WHITE_TERRACOTTA_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> ORANGE_TERRACOTTA_WALL = ITEM_REGISTRY.register("orange_terracotta_wall", () -> new BlockItem(ModBlocks.ORANGE_TERRACOTTA_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> MAGENTA_TERRACOTTA_WALL = ITEM_REGISTRY.register("magenta_terracotta_wall", () -> new BlockItem(ModBlocks.MAGENTA_TERRACOTTA_WALL.get(), PROPERTIES));
@@ -202,6 +219,7 @@ public class ModItems {
     public static final RegistryObject<Item> GREEN_GLAZED_TERRACOTTA_WALL = ITEM_REGISTRY.register("green_glazed_terracotta_wall", () -> new BlockItem(ModBlocks.GREEN_GLAZED_TERRACOTTA_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> RED_GLAZED_TERRACOTTA_WALL = ITEM_REGISTRY.register("red_glazed_terracotta_wall", () -> new BlockItem(ModBlocks.RED_GLAZED_TERRACOTTA_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> BLACK_GLAZED_TERRACOTTA_WALL = ITEM_REGISTRY.register("black_glazed_terracotta_wall", () -> new BlockItem(ModBlocks.BLACK_GLAZED_TERRACOTTA_WALL.get(), PROPERTIES));
+    // CONCRETE WALLS
     public static final RegistryObject<Item> WHITE_CONCRETE_WALL = ITEM_REGISTRY.register("white_concrete_wall", () -> new BlockItem(ModBlocks.WHITE_CONCRETE_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> ORANGE_CONCRETE_WALL = ITEM_REGISTRY.register("orange_concrete_wall", () -> new BlockItem(ModBlocks.ORANGE_CONCRETE_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> MAGENTA_CONCRETE_WALL = ITEM_REGISTRY.register("magenta_concrete_wall", () -> new BlockItem(ModBlocks.MAGENTA_CONCRETE_WALL.get(), PROPERTIES));
@@ -218,6 +236,7 @@ public class ModItems {
     public static final RegistryObject<Item> GREEN_CONCRETE_WALL = ITEM_REGISTRY.register("green_concrete_wall", () -> new BlockItem(ModBlocks.GREEN_CONCRETE_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> RED_CONCRETE_WALL = ITEM_REGISTRY.register("red_concrete_wall", () -> new BlockItem(ModBlocks.RED_CONCRETE_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> BLACK_CONCRETE_WALL = ITEM_REGISTRY.register("black_concrete_wall", () -> new BlockItem(ModBlocks.BLACK_CONCRETE_WALL.get(), PROPERTIES));
+    // WOOL WALLS
     public static final RegistryObject<Item> WHITE_WOOL_WALL = ITEM_REGISTRY.register("white_wool_wall", () -> new BlockItem(ModBlocks.WHITE_WOOL_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> ORANGE_WOOL_WALL = ITEM_REGISTRY.register("orange_wool_wall", () -> new BlockItem(ModBlocks.ORANGE_WOOL_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> MAGENTA_WOOL_WALL = ITEM_REGISTRY.register("magenta_wool_wall", () -> new BlockItem(ModBlocks.MAGENTA_WOOL_WALL.get(), PROPERTIES));
@@ -234,6 +253,7 @@ public class ModItems {
     public static final RegistryObject<Item> GREEN_WOOL_WALL = ITEM_REGISTRY.register("green_wool_wall", () -> new BlockItem(ModBlocks.GREEN_WOOL_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> RED_WOOL_WALL = ITEM_REGISTRY.register("red_wool_wall", () -> new BlockItem(ModBlocks.RED_WOOL_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> BLACK_WOOL_WALL = ITEM_REGISTRY.register("black_wool_wall", () -> new BlockItem(ModBlocks.BLACK_WOOL_WALL.get(), PROPERTIES));
+    // METAL WALLS
     public static final RegistryObject<Item> IRON_BLOCK_WALL = ITEM_REGISTRY.register("iron_block_wall", () -> new BlockItem(ModBlocks.IRON_BLOCK_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> GOLD_BLOCK_WALL = ITEM_REGISTRY.register("gold_block_wall", () -> new BlockItem(ModBlocks.GOLD_BLOCK_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> DIAMOND_BLOCK_WALL = ITEM_REGISTRY.register("diamond_block_wall", () -> new BlockItem(ModBlocks.DIAMOND_BLOCK_WALL.get(), PROPERTIES));
@@ -241,6 +261,7 @@ public class ModItems {
     public static final RegistryObject<Item> NETHERITE_BLOCK_WALL = ITEM_REGISTRY.register("netherite_block_wall", () -> new BlockItem(ModBlocks.NETHERITE_BLOCK_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> LAPIS_BLOCK_WALL = ITEM_REGISTRY.register("lapis_block_wall", () -> new BlockItem(ModBlocks.LAPIS_BLOCK_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> EMERALD_BLOCK_WALL = ITEM_REGISTRY.register("emerald_block_wall", () -> new BlockItem(ModBlocks.EMERALD_BLOCK_WALL.get(), PROPERTIES));
+    // STONE WALLS
     public static final RegistryObject<Item> CALCITE_WALL = ITEM_REGISTRY.register("calcite_wall", () -> new BlockItem(ModBlocks.CALCITE_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> TUFF_WALL = ITEM_REGISTRY.register("tuff_wall", () -> new BlockItem(ModBlocks.TUFF_WALL.get(), PROPERTIES));
     public static final RegistryObject<Item> SMOOTH_BASALT_WALL = ITEM_REGISTRY.register("smooth_basalt_wall", () -> new BlockItem(ModBlocks.SMOOTH_BASALT_WALL.get(), PROPERTIES));